Move graphviz rendering to pkg.
